# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación de mi API con Echo.",
        "title": "API de Sliceflow",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1000",
    "basePath": "/",
    "paths": {
...
        "/hornero/authed/stock/history": {
            "get": {
                "description": "Permite filtrar movimientos por SKU, fecha de inicio y fecha de fin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Obtener historial con filtros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU del producto",
                        "name": "sku",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha inicio (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha fin (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.StockMovement"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hornero/authed/stock/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene la lista completa de productos en stock que no han sido borrados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Listar todos los productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.StockItem"
                            }
                        }
                    },
                    "409": {
                        "description": "Error al obtener productos",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hornero/authed/stock/movement": {
            "post": {
                "description": "Genera un movimiento y actualiza el stock automáticamente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Registrar entrada o salida de stock",
                "parameters": [
                    {
                        "description": "Datos del movimiento",
                        "name": "movement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.CreateMovementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Movement created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error de validación o Stock insuficiente",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hornero/authed/stock/movement/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener métricas del negocio",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.DashboardResponse"
                        }
                    }
                }
            }
        },
        "/hornero/authed/stock/product": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra un producto en el catálogo usando el SKU (código de barras) y nombre.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Crear un nuevo producto",
                "parameters": [
                    {
                        "description": "Datos del producto a crear",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.ProductCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product has been created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Json o error de validación",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "El SKU ya existe",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hornero/authed/stock/{sku}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca un producto específico usando su ID numérico de base de datos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Obtener producto por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "sku",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.StockItem"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Producto no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Edita nombre, descripción, status o stock mínimo de un producto existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Actualizar datos de un producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "sku",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.ProductUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error de validación o ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Realiza un soft-delete del producto usando su ID.",
                "tags": [
                    "Stock"
                ],
                "summary": "Baja lógica de producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "sku",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Product has been deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Error al eliminar",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_francotraversa_Sliceflow_internal_types.CreateMovementRequest": {
            "type": "object",
            "required": [
                "quantity",
                "sku",
                "type"
            ],
            "properties": {
                "location_id": {
                    "description": "Opcional por ahora",
                    "type": "integer"
                },
                "quantity": {
                    "description": "Siempre positivo, el Tipo define si suma o resta",
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "type": {
                    "description": "Solo permitimos IN o OUT",
                    "type": "string",
                    "enum": [
                        "IN",
                        "OUT",
                        "RETURN",
                        "LOSS",
                        "SCRAP",
                        "ADJUST"
                    ]
                },
                "user_id": {
                    "description": "Viene del token/contexto",
                    "type": "integer"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.DashboardResponse": {
            "type": "object",
            "properties": {
                "active_users": {
                    "type": "integer"
                },
                "low_stock_count": {
                    "description": "Cuántos productos están en alerta",
                    "type": "integer"
                },
                "low_stock_items": {
                    "description": "La lista de esos productos",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.StockItem"
                    }
                },
                "movements_today": {
                    "type": "integer"
                },
                "top_selling_items": {
                    "description": "Los 5 más vendidos",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.TopProduct"
                    }
                },
                "total_items": {
                    "description": "Cantidad de productos en catálogo",
                    "type": "integer"
                },
                "total_value": {
                    "description": "Cuánta plata hay parada en el depósito",
                    "type": "number"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.ProductCreateRequest": {
            "type": "object",
            "properties": {
                "sku": {
                    "type": "string",
                    "x-order": "1",
                    "example": "7791234567890"
                },
                "Quantity": {
                    "type": "integer",
                    "example": 10
                },
                "description": {
                    "type": "string",
                    "example": "Importación Diciembre"
                },
                "name": {
                    "type": "string",
                    "example": "PLA PRINT A LOT 1KG AZUL"
                },
                "price": {
                    "type": "number",
                    "example": 98.5
                }
            }
        },
        "github_com_francotraversa_Sliceflow_internal_types.ProductUpdateRequest": {
            "type": "object",
            "properties": {
                "sku": {
                    "type": "string",
                    "x-order": "1",
                    "example": "7791234567890"
                },
                "description": {
                    "type": "string",
                    "example": "Importación Diciembre"
                },
                "name": {
                    "type": "string",
                    "example": "PLA PRINT A LOT 1KG AZUL"
                },
                "price": {
                    "type": "number",
                    "example": 98.5
                },
                "quantity": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "github_com_francotraversa_Sliceflow_internal_types.StockItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "min_qty": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_francotraversa_Sliceflow_internal_types.StockMovement": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location_id": {
                    "type": "integer"
                },
                "qty_after": {
                    "description": "Cuánto quedó (110)",
                    "type": "integer"
                },
                "qty_before": {
                    "description": "Cuánto había (100)",
                    "type": "integer"
                },
                "qty_delta": {
                    "description": "Cuánto cambió (+10, -5)",
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "stock_sku": {
                    "type": "string"
                },
                "type": {
                    "description": "IN, OUT, ADJUST",
                    "type": "string"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.TopProduct": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "total_sold": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribí \"Bearer \" seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}